# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Meticulous API",
    "description": "REST API for the Meticulous CI/CD platform",
    "contact": {
      "name": "Meticulous Team"
    },
    "license": {
      "name": "UNLICENSED"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "paths": {
...
    "/api/v1/workspace/stored-secrets": {
      "get": {
        "tags": [
          "stored_secrets"
        ],
        "operationId": "list_workspace_stored_secrets",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pipeline_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scope_level",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated stored-secret metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_WorkspaceStoredSecretListItem"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/v1/workspace/variables": {
      "get": {
        "tags": [
          "variables"
        ],
        "operationId": "list_workspace_variables",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pipeline_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scope_level",
            "in": "query",
            "description": "all | organization | project | pipeline",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated variables across org",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_WorkspaceVariableListItem"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PaginatedResponse_WorkspaceStoredSecretListItem": {
        "type": "object",
        "description": "Response wrapper for paginated results.",
        "required": [
          "data",
          "pagination"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/StoredSecretResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "pipeline_name": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "project_name": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "project_slug": {
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  }
                }
              ]
            },
            "description": "The items in this page."
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta",
            "description": "Pagination metadata."
          }
        }
      },
      "PaginatedResponse_WorkspaceVariableListItem": {
        "type": "object",
        "description": "Response wrapper for paginated results.",
        "required": [
          "data",
          "pagination"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id",
                "project_id",
                "name",
                "scope",
                "is_sensitive",
                "created_at",
                "updated_at",
                "project_name",
                "project_slug"
              ],
              "properties": {
                "created_at": {
                  "type": "string",
                  "format": "date-time"
                },
                "id": {
                  "type": "string"
                },
                "is_sensitive": {
                  "type": "boolean"
                },
                "name": {
                  "type": "string"
                },
                "pipeline_id": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "pipeline_name": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "project_id": {
                  "type": "string"
                },
                "project_name": {
                  "type": "string"
                },
                "project_slug": {
                  "type": "string"
                },
                "scope": {
                  "type": "string"
                },
                "updated_at": {
                  "type": "string",
                  "format": "date-time"
                },
                "value": {
                  "type": [
                    "string",
                    "null"
                  ]
                }
              }
            },
            "description": "The items in this page."
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta",
            "description": "Pagination metadata."
          }
        }
      },
      "PaginationMeta": {
        "type": "object",
        "description": "Pagination metadata included in responses.",
        "required": [
          "has_more",
          "count"
        ],
        "properties": {
          "count": {
            "type": "integer",
            "description": "Number of items in this response.",
            "minimum": 0
          },
          "has_more": {
            "type": "boolean",
            "description": "Whether there are more items."
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ],
            "description": "Cursor for the next page (null if no more pages)."
          }
        }
      },
...
      "StoredSecretResponse": {
        "type": "object",
        "required": [
          "id",
          "path",
          "kind",
          "version",
          "metadata",
          "created_at",
          "updated_at",
          "propagate_to_projects"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "type": "string"
          },
          "metadata": {},
          "path": {
            "type": "string"
          },
          "pipeline_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "project_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "propagate_to_projects": {
            "type": "boolean",
            "description": "Org-wide only: when `false`, not exposed to pipelines or project secret UIs (catalog SCM may still use)."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "version": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    }
  },
...
}